+++ src/conferences/radv/sass/_common.sass
@@ -158,3 +158,8 @@ h4,
 h5,
 h6
   font-family: $got
+
+.price-increase
+  &__button
+    .btn
+      @extend .btn-green
