+++ src/conferences/radv/sass/components/_buttons.sass
@@ -1,4 +1,4 @@
-.basic-btn
+.basic-btn, .btn
   display: inline-flex
   font-size: 13px
   height: 58px
@@ -15,7 +15,7 @@
   cursor: pointer
   min-width: 60px
 
-  &__inner
+  &__inner, span
     position: relative
     width: 100%
     height: 100%
@@ -55,7 +55,7 @@
     outline-width: 3px !important
   &:hover, &:focus-visible
 
-    .basic-btn__inner
+    .basic-btn__inner, span
       transform: translateX(0)
 
   &.btn-transparent
@@ -158,11 +158,30 @@
           width: 50px
           height: 30px
 
-
-
     &:hover
       .basic-btn__inner
         transform: translateX(0)
 
+.btn
+  span
+    &:after,
+    &:before
+      content: url('data:image/svg+xml;utf8,<svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px" width="25.4px" height="11.4px" viewBox="0 0 25.4 11.4" style="enable-background:new 0 0 25.4 11.4;" xml:space="preserve"><polygon fill="%23ffffff" points="19.7,0 18.3,1.4 21.6,4.7 0,4.7 0,6.7 21.6,6.7 18.3,10 19.7,11.4 25.4,5.7"/></svg>')
+      position: absolute
+      left: 0
+      top: 0
+      bottom: 0
+      display: flex
+      padding-top: 3px
+      width: 60px
+      height: 100%
+      align-items: center
+      justify-content: center
+    &:after
+      left: auto
+      right: -60px
+  &.btn-green span
+    &:before, &:after
+      content: url('data:image/svg+xml;utf8,<svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px" width="25.4px" height="11.4px" viewBox="0 0 25.4 11.4" style="enable-background:new 0 0 25.4 11.4;" xml:space="preserve"><polygon fill="%23000000" points="19.7,0 18.3,1.4 21.6,4.7 0,4.7 0,6.7 21.6,6.7 18.3,10 19.7,11.4 25.4,5.7"/></svg>')
 .is-hidden
   display: none
