+++ src/conferences/radv/sass/components/_intro.sass
@@ -34,7 +34,6 @@
         .intro__octo
           margin-top: 240px
 
-
   &__octo
     position: absolute
     left: 50%
@@ -475,15 +474,15 @@
   background: url(../img/decor_clock.png) center center no-repeat
   background-size: cover
 
-.timer
-  font-size: 36px
-  font-family: $rb
-  width: 250px
-  display: flex
-  align-items: center
-  justify-content: flex-end
-  position: absolute
-  right: 80px
-  top: 113px
-  height: 76px
-  background: linear-gradient(to right, rgba(255,255,255,0) 0%,rgba(255,255,255,1) 100%)
+// .timer
+//   font-size: 36px
+//   font-family: $rb
+//   width: 250px
+//   display: flex
+//   align-items: center
+//   justify-content: flex-end
+//   position: absolute
+//   right: 80px
+//   top: 113px
+//   height: 76px
+//   background: linear-gradient(to right, rgba(255,255,255,0) 0%,rgba(255,255,255,1) 100%)
