+++ src/conferences/radv/sass/helpers/_mixins.sass
@@ -45,11 +45,11 @@
     @content
 
 =xs
-  @media only screen and (max-width: 480px)
+  @media only screen and (max-width: 599px)
     @content
 
 =xxs
-  @media only screen and (max-width: 320px)
+  @media only screen and (max-width: 480px)
     @content
 
 =xxxs
