+++ src/conferences/radv/sass/helpers/_variables.sass
@@ -1,6 +1,9 @@
 
 \:root
   --accentColor: #24fc89
+  --price-increase-bg: #171717
+  --price-increase-color: #fff
+  --price-increase-border: rgba(255,255,255,0.25)
 
   --checkout-section-bg: #fff
   --checkout-section-color: #000
