+++ src/conferences/radv/templates/index.html
@@ -127,5 +127,6 @@ fakeLinks: true
 {% include "parts/_footer.html" %}
 
 {% include "partials/_notice-panel.html" %}
+{% include "partials/_price-increase.html" %}
 
 {% endblock %}
