+++ src/partials/_price-increase.html
@@ -1,11 +1,12 @@
-<div class="price-increase hide js-local-block" data-local-name="priceIncreaseBanner">
+<div id="price-increase" class="price-increase hide js-local-block" data-local-name="priceIncreaseBanner">
   <div class="container">
     <div class="price-increase__wrap">
-      <div id="price-increase-countdown" class="price-increase__items" data-is-in-person="{{inPerson}}"></div>
-      <div class="price-increase__title">Price increase!
-        Save €160 when you register by 24th July</div>
+      <div id="price-increase-countdown" class="price-increase__items" data-timer></div>
+      <div class="price-increase__title" id="price-increase-title"></div>
       <div class="price-increase__button">
-        <a href="#" class="btn">Order now</a>
+        <a href="/checkout" class="btn ">
+          <span>Order now</span>
+        </a>
       </div>
     </div>
   </div>
