+++ src/partials/sass/_price-increase.sass
@@ -4,21 +4,63 @@
   left: 0
   right: 0
   z-index: 20
-  background: var(--mainBg)
-  color: var(--mainColor)
+  background: var(--price-increase-bg)
+  color: var(--price-increase-color)
   padding: 25px 0
+  border-top: 1px solid var(--price-increase-border)
   font-family: GothamPro, Arial, Helvetica, sans-serif
+  +sm
+    padding: 20px 0
+  &__items
+    height: 48px
+
   &__wrap
     display: flex
     align-items: center
-    justify-content: center
     gap: 15px 50px
+    +laptop
+      gap: 15px 35px
+    +md
+      flex-wrap: wrap
+    +sm
+      gap: 7px 0
+
   &__title
-    font-size: 18px
+    font-size: 20px
     font-weight: 500
-    text-align: center
+    line-height: 1.2
+    text-align: left
+    letter-spacing: -0.2px
+    text-transform: uppercase
+    max-width: 580px
+    margin-left: auto
+    margin-right: auto
+
+    +laptop
+      font-size: 18px
+    +md
+      text-wrap: balance
+      text-align: center
+      order: -1
+      br
+        display: none
+    +sm
+      font-size: 16px
+
+    +xxs
+      font-size: 14px
+
   &__button
     flex: none
+    margin-left: auto
+    +md
+      margin-left: 0
+      margin-top: 8px
+    .btn
+      +md
+        min-width: 240px
+      +sm
+        height: 46px
   &__close
     position: absolute
     top: 5px
@@ -34,13 +76,17 @@
       top: 50%
       width: 20px
       height: 1px
-      background: var(--mainColor)
+      background: var(--price-increase-color)
 
       &:nth-child(1)
         transform: translate(-50%, -50%) rotate(45deg)
 
       &:nth-child(2)
         transform: translate(-50%, -50%) rotate(-45deg)
+  .container
+    // for some conf
+    padding-top: 0
+    padding-bottom: 0
 .timer
   display: flex
   gap: 5px
@@ -48,14 +94,19 @@
   &__item
     width: 62px
     text-align: center
+    +md
+      width: 58px
   &__separator
-    border-right: 1px solid var(--mainColor)
-    opacity: 0.15
+    border-right: 1px solid var(--price-increase-color)
+    opacity: 0.1
     height: 32px
     flex: none
   &__nums
     font-size: 26px
     font-weight: 600
-    line-height: 1.2
+    line-height: 32px
+    overflow: hidden
+    +md
+      font-size: 20px
   &__title
     font-size: 12px
